Add `fmt` trait impls for heap containers
This branch adds `fmt::Debug`, `fmt::Display`, and `fmt::Pointer` impls
to `HeapArc`, `HeapArray`, and `HeapFixedVec`, as appropriate. The
`Display` and `Debug` impls forward to the inner types, and require
`T: fmt::Debug` or `T: fmt::Display`; on the other hand, the
`fmt::Pointer` impls are always present regardless of `T` and display
the address of the heap allocation.

I wanted the `fmt::Pointer` impls specifically for tosc-rs/mnemos#4, so
that the bbqueue traces could include the address of the shared buffer,
but they didn't exist at the time, so I thought they would be nice to
add.