melpo: use Tokio to run simulator services (sleeps)
This builds on PR #5, but instead of spawning a separate OS thread for
every `Sleepy` future, we instead change the simulator to run in a
single-threaded Tokio runtime, with each sleep driven by a task spawned
on that runtime. This way, we only need a single thread to drive all of
the simulator services.

In the future, we can also use the same runtime to drive any simulator
services that use I/O (e.g. implementing a simulated serial port using a
host TCP socket). We may also want to use Tokio synchronization
primitives to block the simulator rt on kernel initialization, instead
of the current spin loop, but I'll address that in a separate PR.