feat(melpo): add a simple CLI for configuring the simulator
This branch adds a simple CLI for configuring Melpomene, using `clap`.
Currently, the CLI can be used for configuring options for various
`tracing` backends, and for setting the address on which to bind the TCP
socket for the simulated serial port. More options will probably be
added in the future for configuring other simulator parameters and
debugging features.