fix(melpo): Give CLI args less ambiguous names
Prior to this, both the console addr and modality addr were getting the same default constructor, meaning both the console and modality were trying to bind/connect to port 6699, which seems like a weird and fun behavior of flatten. I've made the names more verbose now to avoid the issue.