Add "ui-test" helper for unit testing (#10)
This PR adds a couple of helpers that allow for writing rustc-style "ui tests", where you can specify a chunk of
forth code as a runnable test. This replaces quite a number of ad-hoc structures to achieve this in the current
unit testing frameworks.

Options are provided for testing blocking, async, or both VM configurations. 

---

* Initial ui tester
* Port the standalone ui-test capability to a testutil
* Docs and more methods
* Ready to review
* Removed standalone ui-test project
* Ported all tests over to runtest infra
* Added async runtest infra
* Remove temp hack
* Switch ordering per suggestion