testutil: print input and output during tests (#13)
The current test utilities are somewhat harder to debug than the
previous `test_lines` functions they replaced, as they do not print
the input and output of the forth VM during the test. This means that
when an assertion failed, it's not clear what input line was executed
immediately before the assertion failed. This branch changes the new
test utils to print each input and output line as the test runs, the way
the previous tests did.

Alternatively, I could change these tests to only print the line of
input immediately before the failure, but I thought it was potentially
useful to get every line...

Inputs and outputs are not printed when running under Miri, for 
performance reasons.