allow forking VMs by refcounting dictionaries (#12)
In mnemOS, we intend to use `forth3` VMs to implement userspace
processes. In order to implement multitasking, it is necessary to
provide a way for one Forth VM to spawn a new child VM as a new process.
In order to do that, it is desirable to have a method of sharing the
contents of a dictionary between VMs.

This branch introduces a notion of shared, reference-counted immutable
dictionaries. A VM's dictionary can be forked by converting its current
dictionary from exclusively owned and mutable to reference counted and
immutable. Then, a new allocation for that VM's dictionary is created,
which points to the frozen dictionary via a "parent" pointer. The forked
VM is also created with a new mutable dictionary which also has the
previous dictionary as its parent. When a definition is not found in a
VM's current dictionary, the parent link is traversed, if it has one.

Parent links hold a reference count on the frozen dictionary, and when
all parent links are dropped, the frozen dictionary is deallocated. This
required changing the API for constructing dictionaries to accept a
callback that is run when a dictionary is deallocated. This allows
gluing dictionary deallocation into an arbitrary user-defined allocation
scheme, which may or may not be compatible with Rust's `GlobalAlloc`
trait (as mnemOS' async allocator cannot implement `GlobalAlloc`).