Initial Forth Shell integration (#45)
This commit adds an initial integration of the Forth VM from `forth3`
into mnemOS, building on the console implementation added in #42 to
display the user's input into the forth VM and the forth VM's output.
The forth VM runs in a separate task, with its "standard in" and
"standard out" streams represented by a bidirectional BBQueue.

Code for running the Forth VM task is in the `kernel` crate, but the
current console implementation is Melpomene-specific. This should be
addressed eventually. Currently, only a single Forth task is spawned,
but subsequent changes will allow a Forth "PID 0" task to spawn child
tasks.

Co-authored-by: James Munns <james@onevariable.com>
Signed-off-by: Eliza Weisman <eliza@elizas-website>