add `Send` and `Sync` impls for `Forth` (#20)
The `Forth` and `AsyncForth` types are `!Send`. This makes them
difficult to use with the latest `maitake`, which requires futures be
`Send` in order to be spawned as a task.

These types are `!Send` because they contain raw pointers to the buffer
regions provided when the VM is constructed. I think they can be made
`Send` (and `Sync`), based on the rationale that the unsafety is in the
invariants of calling `Forth::new` --- if the VM is sent across a thread
boundary, the buffer regions provided to `Forth::new` must travel with
it. Thus, `Forth::new` should still be unsafe, but the VM *can* be
`Send` if `Forth::new`'s invariants are upheld.

Alternatively, we could solve this by wrapping `Forth` in a type which
has ownership over the buffers plus a deallocation callback for those
buffers, instead. This would be similar to `LBForth` but without the use
of `Box`. That type could always be `Send` and `Sync`, without the
caveats I mentioned above.