remove `harness = "false"` from the kernel
currently, the kernel crate doesn't use any custom test harnesses, so
this is not necessary. since the kernel is intended to be cross
platform, it's unclear to me whether we will ever want to write
platform-specific tests in the kernel crate, or if we'll just run host
tests there and run platform-specific tests in platform crates.

in any case, `harness = "false"` was making `cargo test` runs in the whole
workspace fail, so i've removed it. alternatively, we could change CI
to only test specific packages that have host tests, but it's nice for
`cargo test`/`cargo nextest run` to Just Work.

@jamesmunns, does that seem reasonable to you?