chore: run tests on CI (#63)
Currently, CI just runs `cargo check` and `cargo fmt`, and never
actually runs host tests. Some crates in the repo, such as `forth3`,
contain tests that should be run on the host.

This branch adds a `cargo test` job to the CI workflow, as well as a
`cargo miri test` job for `forth3`'s Miri tests.

In order to make this build, I had to remove the `main.rs` from `mstd`,
as it was causing a build error. Since we're not currently using this to
link with a user-provided main function, it's okay to get rid of this
code for now. We can add it back later if we need it. 

Similarly, I also removed `harness = "false"` from the `kernel` crate's
`Cargo.toml`. Currently, the kernel crate doesn't use any custom test
harnesses, so this is not necessary. Since the kernel is intended to be
cross platform, it's unclear to me whether we will ever want to write
platform-specific tests in the kernel crate, or if we'll just run host
tests there and run platform-specific tests in platform crates.

In any case, `harness = "false"` was making `cargo test` runs in the
whole workspace fail, so I've removed it. Alternatively, we could change
CI to only test specific packages that have host tests, but it's nice
for `cargo test`/`cargo nextest run` to Just Work.

Closes #58