feat(d1, crowtty): add UART tracing (#80)
This branch adds an initial implementation of a `tracing` collector that
writes traces to virtual serial port 3 in a COBS+`postcard` binary
format. In addition, it adds a trace viewer capability to `crowtty` when
demuxing port 3, so that the traces written to port 3 are displayed in
`crowtty`'s output. The previous output from `crowtty` that records
every chunk of bytes read/written is now only displayed when a
`--verbose` argument is passed.

Normal `crowtty` output: 

![image](https://github.com/tosc-rs/mnemos/assets/2796466/45d2deb1-27fa-4bbd-a80d-2a43165b094a)

Verbose mode:
![image](https://github.com/tosc-rs/mnemos/assets/2796466/abfbcb75-bd35-43c0-9c3b-6c11146cc664)