feat(melpo): trace improvement (#99)
This commit demotes the traces emitted by Melpomene on every iteration
of the kernel run loop from DEBUG to TRACE, as these occur *extremely*
frequently and spam the console when trying to get debugging output from
other parts of the kernel. In addition, I've added new tracing messages
when a kernel server is started in Melpomene so that we can easily trace
the boot process.