feat(d1): Add sharp display driver (#114)
This restores/extracts the experimental display driver from #110 as a
replacement to the "blinky pattern" that currently exists. This driver
is largely based on the `emb_display` driver that exists in melpo.

The driver is still not *great*, but we should probably rethink the
`emb_display` in melpo first, and update this driver to be consistent
with that.