fix(crowtty): don't panic on 1-byte SerMux frames (#117)
Currently, `crowtty` calls `carry[..used].split_at(2)` every time it
decodes a COBS-encoded frame SerMux frame, to extract the port component
from the frame. However, if we send a zero byte while not currently in a
frame, the COBS decoder will interpret it as a frame with a single zero
in it. When we call `split_at`, crowtty will panic, because the slice
only has one byte in it.

Commit 52310019c0c1e36d6b83eb57f156275ea144922a changed the panic
handler to always send a zero *before* writing out the panic message, in
order to terminate the current SerMux frame if one is in the process of
being sent. However, that zero makes crowtty crash if it is *not*
decoding a partial frame. This means that, rather than nicely printing
the panic message, crowtty just panics immediately when the board
panics. That kinda sucks.

This branch fixes this, by changing crowtty to check whether the length
of the decoded COBS frame is less than the minimum frame length (the
2-byte port number). Now, if we decode a single-byte zero frame, we'll
just ignore it and keep going.