feat(trace): print some identifying information
This commit changes the kernel's `tracing` subscriber so that when a
host crowtty connects and selects a tracing level, the kernel
immediately prints identifying information, including the version, Git
commit the kernel was built from, target, and rustc profile.

This serves two purposes: one, it lets the user know that the desired
trace level has been selected and the board isn't hung, and it allows
the user to know what version of the kernel they're running. Knowing the
version and Git SHA is useful in deveopment in cases where, for example,
you've previously flashed the board with a different dev branch than the
one you're trying to test.