feat(trace): track dropped data (#126)
The kernel's `tracing` subscriber may sometimes drop data, if the
internal trace buffer is full. I hate dropping data. What I hate even
more than dropping data, though, is *silently* dropping data. Sometimes
you gotta drop stuff, but not letting the user know that they are
getting an incomplete picture is unexcusable, especially in a debugging
tool.

So, this branch updates the `trace-proto` crate to include a message for
discarded data. The kernel's collector counts when it discards
something, and the worker task periodically checks if we've lost data,
and sends the new `Discarded` message to let the host know that it
hasn't gotten the whole trace.

I tested this by artificially setting the buffer size to 8 bytes, and it
definitely works:

```
[3 +0282.589088566s] UART BEAT Max level set to Some(DBUG)
[3 +0285.593108198s] UART Discarded { new_spans: 0, span_activity: 0, events: 0, metas: 4 }
```