fix(sermux): `OwnedPortChunk` leaving port in chunk
Currently, the behavior of the `OwnedPortChunk::decode` function is
incorrect. The returned `chunk` in the `OwnedPortChunk` is constructed
by allocating a new `Vec` from the input, decoding a `PortChunk` from
it, and then shrinking the `Vec` to the size of the `chunk` component of
the returned borrowed `PortChunk`.

This does the wrong thing. `Vec::shrink_to` shrinks the `Vec` from the
*end*, so we are shrinking it down to the *first* `chunk.len()` bytes.
The port number is at the beginning of the `Vec`, rather than the end,
so instead of dropping the port number and keeping the data, we are
dropping the last two bytes of the data and producing a buffer that
contains `[port[0], port[1], data[0], ... data[len(data) - 2]]`. This
means the data is silently corrupted.

This branch fixes this behavior by instead constructing the owned chunk
using `pc.chunk.to_vec()`. This has the unfortunate consequence of
meaning that we allocate twice, rather than once, but it means that we
now return the correct data when using `OwnedPortChunk`. Alternatively,
we could use `Vec::remove` twice to remove the first two bytes of the
`Vec`, but that's two _O_(_n_) operations, which is probably at least as
bad as allocating. We could, potentially, change the
`OwnedPortChunk::decode` operation to not just call `PortChunk::decode`,
but that would be more complicated.

Adding a failing test for this is as simple as adding a round-tripping
test for `OwnedPortChunk` as well as `PortChunk`, which we didn't have
previously. I've also changed the round-tripping tests to use `proptest`
to generate arbitrary port numbers and chunks to round trip,
because...why not?