feat(mq-pro): run the graphical Forth shell
This branch changes the `boards/mq-pro` and `boards/lichee-rv` binary
targets to run the graphical Forth shell on its SHARP Memory Display. I
added a utility function in the `d1-core` crate for spawning both the
display driver and the graphical Forth shell, but I decided that the
individual board targets should be responsible for actually calling that
function. My thinking was that the `d1-core` crate is SOC-specific but
not hardware specific, so (theoretically) it knows that it's running on
a D1 but it doesn't know what hardware is connected to that D1.

I also changed `d1-core`'s `D1::initialize` function to spawn the Forth
spawnulator, because...we need to ensure that it's spawnulated so that
we can spawnulate. :)

```
<| ok
```