feat(crowtty): basic built-in pseudo-keyboard (#135)
Currently, the SerMux pseudo-keyboard port is handled like any other
SerMux port by crowtty --- so, if we want to send keyboard input to the
target device, we must open a separate `stty`/`netcat` session to write
to the pseudo-keyboard. This is a bit annyoing and requires opening an
additional terminal window.

To improve developer quality of life, I've added a very basic built-in
pseudo-keyboard to crowtty. Now, input from stdin is written to the
pseudo-keyboard port by default. A command-line argument allows
disabling this behavior and treating the pseudo-keyboard port as a
normal SerMux port, so the old behavior is still available.

The current pseudo-keyboard is pretty basic. It sends input to the
target line-by-line rather than character-by-character, so we can't see
typing on the target's display. Also, everything written to the port is
displayed on the terminal on the host, which is a bit ugly when it gets
interrupted by a log line. In the future, we may want to use a terminal
UI library to change the behavior for stdin to read
character-by-character and not echo the input, so that the input is
instead echoed on the target's display as the user types it in. But this
works for now.

Also, I fixed a typo in the `sermux_proto::WellKnown` enum --- it's
spelt "pseudo", not "psuedo". :)

For example: 

![image](https://github.com/tosc-rs/mnemos/assets/2796466/466dfe45-ec8c-4b74-ae66-4949be942702)

And on the target:

![image](https://github.com/tosc-rs/mnemos/assets/2796466/818701ef-a71a-4de4-9daa-ffa25f42a725)