feat(d1): implement I2C driver (#103)
## Motivation

The Beepy keyboard is driven by firmware (called `i2c_puppet`) running
on an RP2040 MCU, which can be communicated with over the I²C
bus. In order to use the keyboard, we therefore must implement a driver
for the Allwinner D1's I²C controller.

Closes #101
Closes #102

## Solution

This branch adds a definition for an I²C driver service to
`kernel::services`, and an implementation of the I²C driver
service for the Allwinner D1's TWI hardware.

### I²C Service

The cross-platform kernel cannot implement the server for the
I²C driver, because the driver implementation is
platform-specific; but we can provide an abstraction for interfacing
with drivers for a platform's I²C hardware. The
I²C service provides an `I2cClient` type, which is used to
initiate bus operations an I²C target device at a provided
address. Once a bus transaction is initiated, a `Transaction` type is
performed, which can be used to perform multiple read and write
operations with the target device. When the `Transaction` is completed,
a `STOP` condition is sent on the bus; until then, completing a read or
write will instead send a repeated-`START`. This interface permits the
common I²C pattern of writing a register address to read
from, and then reading the contents of that register from the target
within the same transaction.

In addition, the transaction interface can be used to implement the
`embedded_hal_async` crate's `I2c` trait for the `I2cClient` type. This
provides a higher level interface for I²C operations, and
allows our I²C driver to be used with third-party libraries
implemented against the `embedded_hal_async` interface.

Because of mnemOS' message-passing design, the `I2cService` operates
with owned buffers, rather than borrowed buffers, so a `FixedVec<u8>` is
used as the buffer type for both read and write operations. This means
that we must allocate when performing I²C operations. To
reduce the amount of allocation necessary, all `Transaction` methods
return the buffer that was passed in, allowing the buffer to be reused
for multiple operations. To facilitate this, the `Transaction::read` and
`Transaction::write` methods also take a `len` parameter indicating the
actual number of bytes to write from the buffer/read into the buffer,
rather than always writing the entire buffer contents or filling the
entire buffer with bytes. This way, we can size the buffer to the
largest buffer required for a sequence of operations, but perform
smaller reads and writes using the same `FixedVec<u8>`, avoiding
reallocations. The implementation of
`embedded_hal_async::i2c::I2c::transaction` will allocate a single
buffer large enough for the largest operation in the transaction, and
reuse that buffer for every operation.

### TWI driver

This branch adds an implementation of a driver for controlling the
I²C hardware on the D1, which the D1 manual calls a TWI
(Two-Wire Interface), likely due to I²C being a trademark of
Phillips Semiconductor. The TWI hardware can be used in one of two
modes: "TWI engine" mode, where individual bytes are written/read from
the I²C bus in an interrupt handler, and "TWI driver" mode,
where the TWI hardware can operate at the level of I²C
register read/writes using a DMA buffer. This branch only implements a
driver for the TWI engine mode, since it can model all forms of
I²C operations. In the future, we will likely want to
opportunistically use the offload capabilities of the TWI driver when
the I²C transaction has the correct shape for offloading, but
this branch just implements the simpler TWI engine mode.

The TWI hardware is a bit difficult to use correctly, so implementing
this was a bit of a struggle. In particular, it turns out that the
generation of I²C clock pulses occurs when the `TWI_CNTR`
register, which controls the TWI, is written to.

The driver works by sharing state between a driver task and an ISR,
since the TWI engine mode is interrupt driven. The shared state is
"locked" by disabling TWI interrupts temporarily while the driver task
is writing to the shared state, and resuming interrupts when a write to
the shared state completes. In theory, this driver is also safe for use
on multi-core hardware, since a single daemon task is responsible for
all the non-ISR writes to this shared state, but this doesn't actually
matter, because the D1 is an inherently single-core CPU.

The D1 has four separate TWI controllers, `TWI0`, `TWI1`, `TWI2`, and
`TWI3`. The pin mapping for the 40-pin Pi header `I2C0` pins differs
between the MangoPi MQ Pro and the Lichee RV Dock. On the MQ Pro, `TWI0`
is used, with `SCL` on pin `PG12` and `SDA` on pin `PG13`. On the Lichee
RV Dock, `TWI2` is used instead, with `SCL` on pin `PB0` and `SDA` on
pin `PB1`. Therefore, the respective pin mappings are set up by the
individual board support targets, rather than in `mnemos_d1_core`.

### Validation

I've validated the implementation by testing with the Beepy's
`i2c_puppet` device. Now, when running on the Beepy, we spawn a task
that performs a simple read operation to get `i2c_puppet`'s firmware
version, and then sends commands to `i2c_puppet` to turn the Beepy's RGB
LED green. This works :)


![image](https://github.com/tosc-rs/mnemos/assets/2796466/8d8d452a-ce51-4000-8e96-53e1f78deb01)

![image](https://github.com/tosc-rs/mnemos/assets/2796466/c0c83c2e-95c8-4141-aeb4-42de95ae1ae0)

## Future Work

Some things this branch *doesn't* do, but that we will want to do in the
future, include:

- [ ] Implement support for the "TWI driver" DMA offload mode, as
  discussed above
- [ ] Implement 10-bit I²C addresses. The current
  implementation only supports 7-bit address mode. (#143)
- [ ] Potentially, add a driver for MnemOS operating as the
  I²C bus target device rather than as the bus controller?
  I'm not sure if this is actually something we would ever need,
  though...
- [ ] Implement a proper driver for `i2c_puppet`, so that we can use the
  keyboard.
- [ ] A potential optimization involving an `I2cClient` reusing the same
allocated buffer across multiple `embedded_hal_async`
`I2c::transaction`s  could be worth implementing.