fix(d1): fix wrong TWI `A_ACK` flag behavior (#152)
The `A_ACK` bit is only supposed to be set when reading bytes from the
TWI when operating as the bus controller. And, it's supposed to be
disabled *before* reading the final byte of the read, so that the last
byte gets NAKed. We weren't doing that last thing correctly for
multi-byte reads, due to a wrong comparison (sorry). This branch fixes
it.

In addition, I've removed some unneeded setting of `A_ACK` in a few
places to ensure it's only set when it's supposed to be.