add web-based simulator (#148)
this PR adds a web-based (WASM) simulator, codenamed "pomelo". 
It features an interactive shell which allows you to
- start (but not stop) the hello service
- interact with the loopback service via the `echo` command
- interact with a forth process using `forth <single expression>` or
`iforth` for a REPL.

The shell includes a command history that can be navigated with cursor
up/down, and (very) basic tab completion.

Pomelo harnesses `wasm-tracing` to send tracing messages to the browser
debug console. Chrome can further use the tracing spans to show a
profiling graph.

---------

Co-authored-by: Anatol Ulrich <e+github@mail.taugt.net>
Co-authored-by: James Munns <james@onevariable.com>