feat(crowtty): add host-side trace filtering (#156)
Currently, our ability to filter traces is limited to setting a max
level on the target. This is not ideal --- sometimes, we may want
`TRACE` events from one module, but enabling `TRACE` globally results in
an unreadably verbose firehose log.

In the rest of the `tracing` ecosystem, this is generally solved by
using *filters*, typically setting a separate maximum level based on
trace targets (module paths). We can't easily implement dynamic target
filtering in the kernel itself, because `mnemos_alloc` doesn't currently
provide us a nice way of working with owned strings for mudle paths.

However, we *can* easily implement this kind of filtering on the host
side, in `crowtty`. This branch adds support for module-path-based
filtering in `crowtty`, controlled by a `--trace` flag, which replaces
the previous `--trace-level`. I felt like breaking the CLI arguments was
probably fine but if someone really cares about backwards compatibility,
we could make the `--trace-level` flag an alias for `--trace <LEVEL>`.

In the future, when we have the capacity to do nice string handling in
the kernel, it would be preferable to have these trace filter
configurations be sent to the target device, and have it perform
filtering. That would allow us to actually skip recording unwanted
traces, reducing the performance impact of turning on more verbose
traces. However, with the current design, we will still send the max
level to the target, so if we enable `INFO` globally and `DEBUG` in some
specific module, we will, at least, still disable the `TRACE` level on
the target.

When the capacity to perform filtering in the target is added, the
existing flag can be kept, and the behavior modified to send a
traceproto message that selects that filter configuration.

For example, here's a `crowtty` session where I enable the `TRACE` level
for the TWI driver, but `INFO` for all other modules. Note that none of
the traces from other parts of mnemOS, such as `comms::bbq` or the SHARP
display driver, are displayed, so we only see the very verbose
diagnostics from the TWI driver.

![image](https://github.com/tosc-rs/mnemos/assets/2796466/fd22347b-820e-4d74-80a4-639050dc2758)

Fixes #153