qol(workspace): Remove alloc patch (#157)
This PR removes the workspace level patch for mnemos-alloc. This is no
longer needed as the crate is now vendored into the monorepo.

The kernel was already indirectly updated to use a path dep in #148 due
to workspace issues, this change furthers that process and updates
`melpo` and `mstd` to do the same.