feat(beepy): add a mediocre `i2c_puppet` driver (#155)
Depends on #154.

This branch adds a simple driver for the `i2c_puppet` firmware on the
Beepy RP2040. It allows controlling the status LED and reading
keypresses.

Key input can be read by sending a message to the driver service to
subscribe to keypresses. The driver task tracks a list of subscriptions
(channels) and broadcasts each keystroke to all subscribers. In
addition, the status LED and backlight can be controlled by sending
messages to the driver task.

The keyboard driver is in a new `mnemos-beepy` crate in `platforms` that
contains drivers for the Beepy hardware that are implemented using
cross-platform driver service definitions. Right now, this is just the
`i2c_puppet` driver, which is implemented against the kernel's
cross-platform `I2cService` interface. However, in the future, if we can
write a generic driver interface for SPI in the kernel crate, we could
move the SHARP display driver into `mnemos-beepy` as well.

Currently, the way the keyboard is implemented is pretty bad: we poll
the keyboard for new keystrokes periodically, as long as at least one
task is subscribed. This is wildly inefficient and we should instead
have the `i2c_puppet` send us an interrupt when new keystrokes are
ready. However, this will require code for configuring the D1 to
actually generate interrupts on GPIO pins, which we haven't written yet.
I'll add this in a later branch and rewrite the keyboard driver to use
that.

Keyboard input isn't actually wired up to anything yet --- we just run a
task that logs all keypresses. A subsequent branch will actually make
the forth shell handle keyboard input.