feat(kernel): add `initialize_default_services`
One issue with the design of kernel services is that the binding between
services and their clients is not enforced at compile time. This means
that it's possible to add new cross-platform services which should be
spawned on all hardware platforms without the hardware platforms failing
to compile if those services are not spawned.

In order to avoid issues where we forget to spawn a cross-platform
service on a particular platform, such as #164, this branch adds a new
`Kernel::initialize_default_services` that spapwns all the default
cross-platform services and daemon tasks. Rather than having to
duplicate the code to spawn these services in each platform
implementation, we can now simple call this method in the platform impl.
New services which we expect to always be spawned can simply be added
here in the cross-platform kernel, and we don't have to remember to add
code for spawning them to each platform.

To allow the platform to customize settings for the default services,
this method takes a struct containing all the services'
`${SERVICE}Settings` types. In the future, that will probably be
replaced by #160.

While working on this, I also did a bit of cleanup/refactoring of the
various services' registration methods and `Settings` types to make them
a bit more consistent.

Closes #164