fix(trace): always reregister when crowtty connects
Currently, we only re-register all `tracing` callsites if we receive a
host trace level request *and* the level has changed. This means that if
crowtty disconnects and then a new crowtty session connects, and the new
session requests the same max level as the previous one, we won't
register the callsites again. This means that the new session won't be
able to understand any trace spans/events we have previously sent
metadata for.

This branch removes the check that the level has changed, and always
blows away the cache. This causes `tracing` to re-register all callsites
every time a new `crowtty` session connects, so that the new session can
actually understand all trace events.